removed FriendPost and YourPost components, renamed Reply to Post
