--- client/src/components/FriendPost/style.css
@@ -1,8 +1,0 @@
-.friend-bubble {
-    border-radius: 35px;
-    gap: 10px;
-}
-
-.friend-bubble-footer {
-    justify-content: flex-end;
-}
