--- client/src/components/YourPost/style.css
@@ -1,30 +1,0 @@
-.your-bubble {
-    border-radius: 35px;
-    gap: 10px;
-}
-
-.reply-button {
-    font-family: var(--main-text-font);
-    width: 100px;
-    margin: 5px;
-}
-
-.name-container {
-    width: 40%;
-    padding: 5px;
-}
-
-.name-container h2 {
-    font-size: 1.1em;
-}
-
-.content-container {
-    display: flex;
-    justify-content: space-between;
-    text-align: left;
-    width: 60%;
-}
-
-.bubble-text {
-    padding: 0;
-}
