+++ client/src/pages/Home/index.jsx
@@ -1,11 +1,9 @@
-import YourPost from "../../components/YourPost";
-import FriendPost from "../../components/FriendPost";
 import './style.css';
 import Auth from '../../utils/auth'
 import { QUERY_USER, QUERY_POSTS, QUERY_USER_INFO } from '../../utils/queries';
 import { DELETE_POST, DELETE_REPLY } from "../../utils/mutations";
 import { useQuery, useMutation } from '@apollo/client';
-import Reply from "../../components/Reply";
+import Reply from "../../components/Post";
 import { useState, useEffect } from 'react'
 
 export default function Home() {
